Add base Ability class that will provides authorization
This class should be overriden in an application that would like to define its own CanCan ability class
