+++ app/controllers/forem/forums_controller.rb
@@ -12,10 +12,5 @@ module Forem
       @topics = forem_admin? ? @forum.topics : @forum.topics.visible
       @topics = @topics.by_pinned_or_most_recent_post.page(params[:page]).per(20)
     end
-
-    private
-      def current_ability
-        Forem::Ability.new(forem_user)
-      end
   end
 end
