+++ app/models/ability.rb
@@ -1,0 +1,8 @@
+# This class should be overriden by your application if you're using CanCan.
+class Ability
+  def initialize(user)
+  end
+
+end
+
+Ability.send :include, Forem::Ability
